infra-tests/requirements.txt
-i https://pypi.org/simple
ansible==2.9.5
apipkg==1.5
appdirs==1.4.3
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
dogpile.cache==0.9.0
execnet==1.7.1
idna==2.9
importlib-metadata==1.5.0 ; python_version < '3.8'
iso8601==0.1.12
jinja2==2.11.1
jmespath==0.9.5
jsonpatch==1.25
jsonpointer==2.0
keystoneauth1==3.18.0
lxml==4.5.0
markupsafe==1.1.1
more-itertools==8.2.0
munch==2.5.0
netifaces==0.10.9
openstacksdk==0.41.0
os-service-types==1.7.0
packaging==20.1
pbr==5.4.4
pluggy==0.13.1
py==1.8.1
pycparser==2.19
pyparsing==2.4.6
pytest-forked==1.1.3
pytest-html==2.0.1
pytest-json-report==1.2.1
pytest-metadata==1.8.0
pytest-sugar==0.9.2
pytest-xdist==1.31.0
pytest==5.3.5
pyyaml==5.3
requests==2.23.0
requestsexceptions==1.4.0
six==1.14.0
stevedore==1.32.0
termcolor==1.1.0
testinfra==4.1.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.0.0
